requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.2
plotly==5.18.0
pandas==2.1.3
numpy==1.26.2
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
aiohttp==3.9.1
requests==2.31.0
pydantic==2.5.2
python-dotenv==1.0.0
watchdog==3.0.0
streamlit-autorefresh
psycopg2-binary==2.9.9
